+++ docs/CHANGELOG.md
@@ -3,10 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Version [0.5.0]
+## Version [0.5.2]
+- [x] Updated documentation with improved plan and steps.
+- [x] Fixed port randomly changing on `npm start`
 - [x] Improved voice respond to text and audio interrupt.
 - [x] manifest.json file for proper package distribution
 - [x] Improved installation instructions and docs (Tutor, etc.)
+- [x] Fixed bug where assistant messages were not merged correctly.
 
 ## Version [0.4.0]
 - [x] Add Listening phrase "Start a Voice" to start voice chat.